Patches by Mark G. Adams for 64 bit linux.

+++ Unicode/ucharmap.c
@@ -396,7 +396,7 @@ return( true );
 unichar_t *def2u_strncpy(unichar_t *uto, const char *from, int n) {
 #if HAVE_ICONV_H
     if ( my_iconv_setup() ) {
-	int in_left = n, out_left = 2*n;
+	size_t in_left = n, out_left = 2*n;
 	char *cto = (char *) uto;
 	iconv(to_unicode, (char **) &from, &in_left, &cto, &out_left);
 	if ( cto<((char *) uto)+2*n) *cto++ = '\0';
@@ -410,7 +410,7 @@ return( encoding2u_strncpy(uto,from,n,local_encoding));
 char *u2def_strncpy(char *to, const unichar_t *ufrom, int n) {
 #if HAVE_ICONV_H
     if ( my_iconv_setup() ) {
-	int in_left = 2*n, out_left = n;
+	size_t in_left = 2*n, out_left = n;
 	char *cfrom = (char *) ufrom, *cto=to;
 	iconv(from_unicode, &cfrom, &in_left, &cto, &out_left);
 	if ( cto<to+n ) *cto++ = '\0';
@@ -431,7 +431,7 @@ return( NULL );
     uto = galloc((len+1)*sizeof(unichar_t));
 #if HAVE_ICONV_H
     if ( my_iconv_setup() ) {
-	int in_left = len, out_left = 2*len;
+	size_t in_left = len, out_left = 2*len;
 	char *cto = (char *) uto;
 	iconv(to_unicode, (char **) &from, &in_left, &cto, &out_left);
 	*cto++ = '\0';
@@ -456,7 +456,7 @@ return( NULL );
     len = u_strlen(ufrom);
 #if HAVE_ICONV_H
     if ( my_iconv_setup() ) {
-	int in_left = 2*len, out_left = 3*len;
+	size_t in_left = 2*len, out_left = 3*len;
 	char *cfrom = (char *) ufrom, *cto;
 	cto = to = galloc(3*len+2);
 	iconv(from_unicode, &cfrom, &in_left, &cto, &out_left);
